fix/java: fix Java local refs leak
When we convert Java objects in a loop, a new ref created with
each iteration. We don't really need these refs because they are
stored as array elements, so now we delete them preemptively to
fix the reference leakage which is important for older Android
JVMs which have a fixed limit of 512 local references.