refactor/result: macros to result.rs, introduce `NativeResult` type
There didn't seem to be a strong reason for using macros over regular functions.
Some refactoring was done to convert the macros into regular functions.

This commit also introduces NativeResult (which ideally would be in its own
commit, but it was hard for me to separate the changes). This struct can be
created with a regular String description and converted into FfiResult, which
now owns its pointer, unlike before. Because FfiResult owns its pointer, it can
clean up the pointed-to data on Drop. This makes it safer to create an
FfiResult, plus it follows our existing conventions for creating FFI
types (start with the native type, then convert), and less mental overhead is
needed when dealing with this type.